Extract U.getTaskbarPosition to TaskbarPosition
